# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export type AuthResponseUser = {
  id: string;
  /** @nullable */
  email?: string | null;
};

export type ProfileRole = typeof ProfileRole[keyof typeof ProfileRole];


export const ProfileRole = {
  customer: 'customer',
  manager: 'manager',
  admin: 'admin',
} as const;

export interface Profile {
  id: string;
  email: string;
  role: ProfileRole;
}

export interface AuthResponse {
  access_token: string;
  refresh_token: string;
  /** @nullable */
  expires_at?: number | null;
  user: AuthResponseUser;
  profile: Profile;
}

export interface Category {
  id: string;
  name: string;
  slug: string;
}

export interface Product {
  id: string;
  category_id: string;
  name: string;
  price: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image_url?: string | null;
  stock_count: number;
  is_available: boolean;
}

export type ListProductsParams = {
category?: string;
};