# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Supabase Auth operations
  - name: menu
    description: Public menu operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Create a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials or Supabase error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: logIn
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logOut
      tags: [auth]
      summary: Sign out the current session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Signed out
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentProfile
      tags: [auth]
      summary: Get the signed-in user's profile and role
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [menu]
      summary: List menu categories
      responses:
        "200":
          description: Menu categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /products:
    get:
      operationId: listProducts
      tags: [menu]
      summary: List available products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Available menu products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthCredentials:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: ["number", "null"]
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: ["string", "null"]
          required:
            - id
        profile:
          $ref: "#/components/schemas/Profile"
      required:
        - access_token
        - refresh_token
        - user
        - profile
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, manager, admin]
      required:
        - id
        - email
        - role
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - slug
    Product:
      type: object
      properties:
        id:
          type: string
        category_id:
          type: string
        name:
          type: string
        price:
          type: number
        description:
          type: ["string", "null"]
        image_url:
          type: ["string", "null"]
        stock_count:
          type: number
        is_available:
          type: boolean
      required:
        - id
        - category_id
        - name
        - price
        - stock_count
        - is_available
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT