# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}