# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResponseUser } from './authResponseUser';
import type { Profile } from './profile';

export interface AuthResponse {
  access_token: string;
  refresh_token: string;
  /** @nullable */
  expires_at?: number | null;
  user: AuthResponseUser;
  profile: Profile;
}