# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  category_id: string;
  name: string;
  price: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image_url?: string | null;
  stock_count: number;
  is_available: boolean;
}