# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileRole } from './profileRole';

export interface Profile {
  id: string;
  email: string;
  role: ProfileRole;
}